Before coding a SVG component
Like with a chocolate calendar, I am making it a habit for myself to keep in touch with SVG and possible applications of these interesting Filter Elements, as they are implemented in a preset found in Inkscape (open-source design tool).

On a side note, I used this one raster that is included within in my project before, one external image that I got online (https://stockarch.com/images/abstract/patterns/red-twisted-mesh-2139, wearing a Creative-Commons tag). I had to change it's colors and I did actually link to it then, but of course I don't know if it's author minds that I colorized it a little. What can I say, I like how the grid is both parallel and bent, giving an impression of a coordinate system, but then again ...

I have shredded paper sketches and drafts, just recently on the balcony, but of course transcribing my thoughts to a digital device (is this a Paper Leap). As do the rainbows pave the roof at the fringe, so does the motive in this picture assert an imaginary horizon, and like so does a live implementation find you on an augmented landscape in a contemporary data grid (bisected by Google Ingress, for example).

(Interchangeable words being "horizon", "skyline" and "sky arc")