Code style update (indentations)
GitHub somehow doesn't do me "Word wrap" or "Line wrap" and syntax highlighting, when that really is a good idea. Maybe there's a "Gamification" goal on the agenda, saying that there should be a trigger that inclines people to either ... a) download or git clone a repository, or b) go design an interface for the contained data.

And also, who made a port of Sublime Text code editor to Android? Just a quick edit, really.