Update seed.jsonion.md
Referring to modification of line 23:

To a web developer who's limited to using JavaScript there's a few things which remain undetailed, without knowing how browsing client software might optimize code execution ... There is Cascading Style Sheets, offering a conventional way to modify visual element's states with custom rules (conditional), which by syntax in brackets looks like a JSON object; however, syntax of path to one or more element nodes doesn't apply to native data objects. A browser functionality is implied; that there is an index of paths to elements that are observed for changes. It is unclear, however, whether any JavaScript code, modifying larger data objects, is prepared ahead and rendered during time of receiving an actual JSON stream; and if path is addressed with the likes of xpath.

Browser clients implement conventional methods, like "Custom Elements" allows for registering custom tags in HTML (> 2018), similar to "xml:ns" and "qnames", which enable use of established namespace.

Code execution as defined in a <script type="text/javascript" /> tag doesn't seem to make use of this built-in index of nodes of visual elements and the path model, so to define events on change to data state, and initialize a data object state observer (defined to match multiple entries by a conventional way, as with CSS).

React for CommonMark is hovering along over the cascades, stacked somewhere in a browser's hood, for all we know.