Create (25.02.2021) #LeapGest.md
Progress has been made since first publishing of this whitepaper with a personal GitHub account. Updates have not yet materialized as updates to the specification that this whitepaper represents... As instead, several partials are now scattered on a professional Facebook account (posted in comments of posts on topic of web development).

There will be a follow-up briefly, implementing the refined logic to tackle further the case of LeapGest (× Leap second). Should you be aware of independent progress on this subject, you are very welcome to submit it hereby, as a comment or a pull request. 

Thank you for your attention, without it evolving this project in collective imagination would've been impossible!

Links in mention:
- https://github.com/blazbrmnn/leapgesture/blob/whitepaper-app/whitepaper/(25.02.2021)%20%23LeapGest.md
- https://facebook.com/blablaz